backend/requirements.txt
# Django Framework & Related
Django==4.2.13
djangorestframework==3.15.2

# API Documentation
drf-yasg==1.21.7

# Authentication & Authorization
djangorestframework-simplejwt==5.5.0
dj-rest-auth==5.1.0
django-allauth==64.2.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3
geopy==2.4.1

# Production Deployment (WSGI Server)
gunicorn==22.0.0

# Test Code
Faker>=18.0.0

# Database Driver (entrypoint.sh에서 동적으로 설치)

# --- Minio/S3 연동을 위해 추가 ---
django-storages
boto3

# --- ImageField 사용을 위해 추가 ---
Pillow